Add bindings for TimePickerAndroid (#63)
Create bindings for TimePickerAndroid based on the bindings for DatePickerAndroid. I used the same conventions used there, but bs.deriving.abstract should probably be used in favor of deprecated . objects. 

Also, I have hours and minutes typed as ints, should I type them as variants? Hour is in fact 0-23 and minute is 0-59.