Exit docs build script if odoc is missing (#313)
The idea is to prevent tons of "Command does not exist" error & point to the right place on the required dependencies.
When they contribute, I guess lots of people (including me) just get a repo, do `yarn && yarn test` to see if everything is fine.
Currently with the repo, you gets tons or error if you blindly do that. This little modification ensure that `odoc` has been installed & warn+exit otherwise.