Remove private modules as it's very annoying to quickly copy/tweak locally a binding that's incomplete
With this change, we don't really have private code anymore, which is unaccessible outside bs-react-native codebase itself.
So if you have to use a module (eg: FlatList) & that a props is missing, you can just copy & paste it, add `open BsReactNative;` at the top & you are good to go to tweak it without having to make a pr or a fork & wait for it to be merged.

The only real "private" part has been moved to an `Internals` module. Hope the name make it clear :)

You are welcome :)

Note: responderUtils seems unused as not exposed rei or used internally, so I just dropped it.