ReactNavigation WIP
This is just a start that going to help us to see where we wanna go.

(Just to let you know, I only used react-navigation once with js, and it was v1 - but I think it didn't changed that much from what I saw on docs)

One question I have: how are we supposed to define screenProps per screen?

I was using js+flow when using react-navigation 1 & here are things that I disliked

- no way to validate that your are going to navigate to a screen that exist (string to navigate)
- no way to know if the screenProps you are going to pass are valid

Not sure how we can address that with reason-react-navigation as I only see on screenProps passed to main functor.

Did I miss something?